+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added dynamic parameters to stack New/Set cmdlets.
 * Used correct JSON serializer settings for all templates-related deserialization.
 
 ## Version 6.16.0